pyproject.toml
[project]
name = "goit-web-hw-13"
version = "0.1.0"
description = ""
authors = [
    {name = "Mykhailo Popudniak",email = "dev.miquella@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
